adjusts normalizeanswer function
