+++ play/package.json
@@ -21,7 +21,7 @@
     "i18next": "^22.4.15",
     "i18next-browser-languagedetector": "^7.0.1",
     "i18next-http-backend": "^2.2.1",
-    "katex": "^0.16.8",
+    "katex": "^0.16.9",
     "mathjs": "^11.11.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
